stt/requirements.txt
faster-whisper==1.0.3
websockets==12.0
numpy==1.26.4
fastapi==0.111.1
groq==0.9.0
deepgram-sdk==3.4.0
python-dotenv==1.0.1
